backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.2.3
--only-binary :all:
plotly==5.18.0
prophet==1.1.4
scikit-learn==1.6.1
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.350
psycopg2-binary==2.9.9
python-multipart==0.0.6 